// src/app/models/coupon-list.ts
export class CouponList {
  CustCouponId: number;
  IssueDate: string;
  ExpDate: string;
  UseDate: string;
  Status: string;
  CType: string;
  Id: number;
  UsageCountLeft: number;
  Name: string;
  Description: string;
  CouponCode: string;
  MaxCap: number;
  Type: string;
  Discount: number;
  PreTax: string;
  Active: string;
  MinOrderAmt: string;
  CouponImage: string;
  Service: string;
  PaymentType: any;
}